requirements.txt
biopython>=1.66
numpy>=1.12.0
scipy>=0.18.1
